+++ src/chartreuse/tests/unit_tests/conftest.py
@@ -12,6 +12,6 @@ def configure_chartreuse_mock(mocker: MockerFixture, is_migration_needed: bool =
     )
     mocker.patch("chartreuse.chartreuse_upgrade.Chartreuse.upgrade")
     mocker.patch(
-        "wiremind_kubernetes.kubernetes_helper._get_namespace_from_kube",
+        "chartreuse.kubernetes_helper._get_namespace_from_kube",
         return_value="foo",
     )